new rows-cols mixin, move circlepacking to overview
