+++ src/pages/index.module.scss
@@ -69,8 +69,3 @@
 		}
 	}
 }
-.circlePacking_col {
-	* {
-		overflow: visible;
-	}
-}
