+++ src/pages/index.tsx
@@ -8,23 +8,16 @@ import breakPoints from "../utils/breakpoints";
 import styles from "./index.module.scss";
 import Label from "../components/Input/Label";
 import DataCountCard from "../components/DataCountCard/DataCountCard";
-import CirclePacking from "../components/CirclePacking/CirclePacking";
 import { useQuery } from "react-query";
-import { getCancerHierarchy, getModelCount } from "../apis/AggregatedData.api";
-import { useRouter } from "next/router";
-import Loader from "../components/Loader/Loader";
+import { getModelCount } from "../apis/AggregatedData.api";
 import SearchBar from "../components/SearchBar/SearchBar";
 
 const Home: NextPage = () => {
 	const { windowWidth } = useWindowDimensions();
 	let bpLarge = breakPoints.large;
-	let cancerHierarchyQuery = useQuery("cancerHierarchy", () => {
-		return getCancerHierarchy();
-	});
 	let modelCount = useQuery("modelCount", () => {
 		return getModelCount();
 	});
-	const router = useRouter();
 
 	return (
 		<>
@@ -75,45 +68,12 @@ const Home: NextPage = () => {
 			</header>
 			<section>
 				<div className="container">
+					<div className="row justify-content-center">
+						<div className="col-10">
+							<DataCountCard layout="horizontal" />
+						</div>
+					</div>
 					<div className="row align-center">
-						<ShowHide windowWidth={windowWidth || 0} showOver={bpLarge}>
-							<div
-								className={`col-12 col-md-10 col-lg-5 offset-md-1 offset-lg-0 ${styles.circlePacking_col}`}
-							>
-								{/* Graph */}
-								<div
-									style={{
-										backgroundColor: "#085154",
-										aspectRatio: "1",
-										borderRadius: "500%",
-									}}
-								>
-									{!cancerHierarchyQuery.isLoading &&
-									cancerHierarchyQuery.data ? (
-										<CirclePacking
-											data={cancerHierarchyQuery.data}
-											onCircleClick={(circleId, circleDepth) => {
-												const searchPrefix =
-													circleDepth === 1
-														? `?filters=cancer_system:`
-														: `?filters=search_terms:`;
-												const termSuffix = circleDepth === 1 ? "Cancer" : "";
-												const search = `${searchPrefix}${encodeURIComponent(
-													circleId + termSuffix
-												)}`;
-
-												router.push({
-													pathname: "/search",
-													search,
-												});
-											}}
-										/>
-									) : (
-										<Loader />
-									)}
-								</div>
-							</div>
-						</ShowHide>
 						<div className="col-12 col-lg-6 offset-lg-1">
 							<h2>
 								We can collect and display your data making you more
@@ -142,7 +102,7 @@ const Home: NextPage = () => {
 					<section className="pt-0">
 						<div className="container">
 							<div className="row">
-								<DataCountCard />
+								<DataCountCard layout="vertical" />
 							</div>
 						</div>
 					</section>
@@ -170,7 +130,7 @@ const Home: NextPage = () => {
 							</div>
 							<ShowHide showOver={bpLarge} windowWidth={windowWidth || 0}>
 								<div className="col-12 col-lg-3 col-xl-2 offset-lg-1 offset-xl-1">
-									<DataCountCard />
+									<DataCountCard layout="vertical" />
 								</div>
 							</ShowHide>
 						</div>
