+++ src/pages/overview.module.scss
@@ -1,0 +1,5 @@
+.circlePacking_col {
+	* {
+		overflow: visible;
+	}
+}
